requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-docx==1.1.0
PyPDF2==3.0.1
pytesseract==0.3.10
Pillow==10.1.0
opencv-python==4.8.1.78
spacy==3.7.2
transformers==4.35.2
torch==2.1.1
pandas==2.1.3
numpy==1.25.2
requests==2.31.0
aiofiles==0.24.0
jinja2==3.1.2
reportlab==4.0.7
openpyxl==3.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
pydantic==2.5.0
pydantic-settings==2.1.0